File Structure Progs Folder
This is the folder that combines all of our files.  On thursday 11/17 we
will discuss how everything will fit into this folder.  Also to use this
folder you need to set up your machine properly, but I am also trying to
figure out how to save a VM so you guys can just upload the VM to
virtual box and use it like I have it.